+++ app/vtinsert/insertutil/common_params.go
@@ -152,6 +152,9 @@ type LogRowsStorage interface {
 
 	// CanWriteData must returns non-nil error if logs cannot be added to the underlying storage.
 	CanWriteData() error
+
+	// IsLocalStorage tells the caller whether the current instance is using local storage.
+	IsLocalStorage() bool
 }
 
 var logRowsStorage LogRowsStorage