+++ app/vtinsert/opentelemetry/otlphttp.go
@@ -75,9 +75,9 @@ func handleProtobufRequest(r *http.Request, w http.ResponseWriter) {
 	encoding := r.Header.Get("Content-Encoding")
 	err = protoparserutil.ReadUncompressedData(r.Body, encoding, maxRequestSize, func(data []byte) error {
 		var callbackErr error
-		lmp := cp.NewLogMessageProcessor("opentelemetry_traces_otlphttp_protobuf", false)
-		callbackErr = pushHTTPProtobufRequest(data, lmp)
-		lmp.MustClose()
+		tsp := cp.NewTraceProcessor("opentelemetry_traces_otlphttp_protobuf", false)
+		callbackErr = pushHTTPProtobufRequest(data, tsp)
+		tsp.MustClose()
 		return callbackErr
 	})
 	if err != nil {
@@ -126,13 +126,13 @@ func handleJSONRequest(r *http.Request, w http.ResponseWriter) {
 			req         otelpb.ExportTraceServiceRequest
 			callbackErr error
 		)
-		lmp := cp.NewLogMessageProcessor("opentelemetry_traces_otlphttp_json", false)
+		tsp := cp.NewTraceProcessor("opentelemetry_traces_otlphttp_json", false)
 		if callbackErr = req.UnmarshalJSONCustom(data); callbackErr != nil {
 			errorsJSONTotal.Inc()
 			return fmt.Errorf("cannot unmarshal request from %d protobuf bytes: %w", len(data), callbackErr)
 		}
-		callbackErr = pushExportTraceServiceRequest(&req, lmp)
-		lmp.MustClose()
+		callbackErr = pushExportTraceServiceRequest(&req, tsp)
+		tsp.MustClose()
 		return callbackErr
 	})
 	if err != nil {