+++ apptest/tests/otlp_ingestion_test.go
@@ -35,6 +35,7 @@ func testOTLPIngestionJaegerQuery(tc *at.TestCase, sut at.VictoriaTracesWriteQue
 	// ingest data via /insert/opentelemetry/v1/traces
 	sut.OTLPHTTPExportTraces(t, req, at.QueryOpts{})
 	sut.ForceFlush(t)
+	time.Sleep(2 * time.Second) // index will be created after -insert.traceMaxDuration (2s in integration test)
 
 	// verify
 	assertFunc()
@@ -62,6 +63,7 @@ func testOTLPgRPCIngestionJaegerQuery(tc *at.TestCase, sut at.VictoriaTracesWrit
 	// ingest data via /insert/opentelemetry/v1/traces
 	sut.OTLPgRPCExportTraces(t, req, at.QueryOpts{})
 	sut.ForceFlush(t)
+	time.Sleep(2 * time.Second) // index will be created after -insert.traceMaxDuration (2s in integration test)
 
 	// verify
 	assertFunc()
@@ -75,7 +77,7 @@ func getDefaultIngestRequestAndAssertFunc(tc *at.TestCase, sut at.VictoriaTraces
 	// prepare test data
 	serviceName := "testKeyIngestQueryService"
 	spanName := "testKeyIngestQuerySpan"
-	traceID := "123456789"
+	traceID := "bda5886e99fffef35a847cb2d493fde0"
 	spanID := "987654321"
 	testTagValue := "testValue"
 	testTag := []*otelpb.KeyValue{
@@ -162,7 +164,7 @@ func getDefaultIngestRequestAndAssertFunc(tc *at.TestCase, sut at.VictoriaTraces
 			Spans: []at.Span{
 				{
 					Duration: 0,
-					TraceID:  hex.EncodeToString([]byte(traceID)),
+					TraceID:  traceID,
 					SpanID:   hex.EncodeToString([]byte(spanID)),
 					Logs: []at.Log{
 						{
@@ -178,7 +180,7 @@ func getDefaultIngestRequestAndAssertFunc(tc *at.TestCase, sut at.VictoriaTraces
 					ProcessID:     "p1",
 					References: []at.Reference{
 						{
-							TraceID: hex.EncodeToString([]byte(traceID)),
+							TraceID: traceID,
 							SpanID:  hex.EncodeToString([]byte(spanID)),
 							RefType: "FOLLOWS_FROM",
 						},
@@ -196,7 +198,7 @@ func getDefaultIngestRequestAndAssertFunc(tc *at.TestCase, sut at.VictoriaTraces
 					},
 				},
 			},
-			TraceID: hex.EncodeToString([]byte(traceID)),
+			TraceID: traceID,
 		},
 	}
 
@@ -252,7 +254,7 @@ func getDefaultIngestRequestAndAssertFunc(tc *at.TestCase, sut at.VictoriaTraces
 		tc.Assert(&at.AssertOptions{
 			Msg: "unexpected /select/jaeger/api/traces/<trace_id> response",
 			Got: func() any {
-				return sut.JaegerAPITrace(t, hex.EncodeToString([]byte(traceID)), at.QueryOpts{})
+				return sut.JaegerAPITrace(t, traceID, at.QueryOpts{})
 			},
 			Want: &at.JaegerAPITraceResponse{
 				Data: expectTraceData,